requirements.txt
alembic==1.14.1
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.44.0
coverage==7.6.1
exceptiongroup==1.2.2
faker==35.2.2
flask==2.3.3
flask-bcrypt==1.0.1
flask-cors==5.0.0
flask-jwt-extended==4.6.0
flask-mail==0.10.0
flask-migrate==4.1.0
flask-restful==0.3.10
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mako==1.3.10
markupsafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.1
packaging==25.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
pluggy==1.5.0
psycopg2-binary==2.9.10
pyjwt==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==8.3.5
pytest-cov==5.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.2
qrcode==7.4.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.10.1
sqlalchemy==2.0.40
sqlalchemy-serializer==1.4.12
starkbank-ecdsa==2.2.0
stripe==12.1.0
tomli==2.0.1
typing-extensions==4.13.2
urllib3==2.2.3
werkzeug==2.3.8
zipp==3.20.2
sendgrid==6.11.0
six==1.17.0
